extract object tab manager
