+++ frontend/app/src/entities/nodes/object-item-details/object-item-details-paginated.tsx
@@ -2,11 +2,11 @@ import {
   DEFAULT_BRANCH_NAME,
   GENERIC_REPOSITORY_KIND,
   MENU_EXCLUDELIST,
-  TASK_TAB,
   TASK_TARGET,
 } from "@/config/constants";
 import { QSP } from "@/config/qsp";
 import { currentBranchAtom } from "@/entities/branches/stores";
+import { NodeEvents } from "@/entities/events/ui/node-details-events";
 import { ObjectAttributeValue } from "@/entities/nodes/getObjectItemDisplayValue";
 import ObjectItemMetaEdit from "@/entities/nodes/object-item-meta-edit/object-item-meta-edit";
 import {
@@ -20,7 +20,6 @@ import {
   RepositoryObjectsTab,
 } from "@/entities/nodes/object/ui/object-tabs";
 import { getRelationshipsVisibleInTab } from "@/entities/nodes/object/utils/get-relationships-visible-in-tab";
-import { ObjectRelationshipsManager } from "@/entities/nodes/relationships/ui/object-relationships-manager";
 import { showMetaEditState } from "@/entities/nodes/stores/metaEditFieldDetails.atom";
 import { metaEditFieldDetailsState } from "@/entities/nodes/stores/showMetaEdit.atom";
 import { NodeObject } from "@/entities/nodes/types";
@@ -29,23 +28,18 @@ import { Permission } from "@/entities/permission/types";
 import { genericSchemasAtom, nodeSchemasAtom } from "@/entities/schema/stores/schema.atom";
 import { ModelSchema, RelationshipSchema } from "@/entities/schema/types";
 import { isGenericSchema } from "@/entities/schema/utils/is-generic-schema";
-import { TaskItemDetails } from "@/entities/tasks/ui/task-item-details";
-import { TaskItems } from "@/entities/tasks/ui/task-items";
 import { queryClient } from "@/shared/api/rest/client";
-import { constructPath } from "@/shared/api/rest/fetch";
 import { ButtonWithTooltip } from "@/shared/components/buttons/button-primitive";
 import MetaDetailsTooltip from "@/shared/components/display/meta-details-tooltips";
 import SlideOver from "@/shared/components/display/slide-over";
 import { Card, CardWithBorder } from "@/shared/components/ui/card";
-import { Link } from "@/shared/components/ui/link";
 import { useTitle } from "@/shared/hooks/useTitle";
 import { LockClosedIcon } from "@heroicons/react/24/outline";
 import { Icon } from "@iconify-icon/react";
 import { useAtom, useAtomValue } from "jotai";
-import { useRef } from "react";
-import { Navigate, useLocation, useParams } from "react-router";
+import { Navigate, useParams } from "react-router";
 import { StringParam, useQueryParam } from "use-query-params";
-import { NodeEvents } from "../../events/ui/node-details-events";
+import { ObjectTabManager } from "../relationships/ui/object-tab-manager";
 import { ActionButtons } from "./action-buttons";
 import { ObjectAttributeRow } from "./object-attribute-row";
 import RelationshipDetails from "./relationship-details-paginated";
@@ -63,12 +57,9 @@ export default function ObjectItemDetails({
   permission,
   hideHeaders,
 }: ObjectDetailsProps) {
-  const location = useLocation();
   const { objectKind, objectid } = useParams();
-  const { pathname } = location;
 
   const [qspTab] = useQueryParam(QSP.TAB, StringParam);
-  const [qspTaskId] = useQueryParam(QSP.TASK_ID, StringParam);
   const [showMetaEditModal, setShowMetaEditModal] = useAtom(showMetaEditState);
   const [metaEditFieldDetails, setMetaEditFieldDetails] = useAtom(metaEditFieldDetailsState);
   const branch = useAtomValue(currentBranchAtom);
@@ -77,8 +68,6 @@ export default function ObjectItemDetails({
   const isTaskTarget = !isGenericSchema(schema) && !!schema.inherit_from?.includes(TASK_TARGET);
   const isRepository =
     !isGenericSchema(schema) && !!schema.inherit_from?.includes(GENERIC_REPOSITORY_KIND);
-
-  const refetchRef = useRef(null);
 
   if ((schemaList?.length || genericList?.length) && !schema) {
     // If there is no schema nor generics, go to home page
@@ -238,35 +227,8 @@ export default function ObjectItemDetails({
         </div>
       )}
 
-      {qspTab && qspTab !== TASK_TAB && (
-        <ObjectRelationshipsManager
-          parentNodeSchema={schema}
-          parentNodeId={objectDetailsData.id}
-          relationshipName={qspTab}
-        />
-      )}
-
-      {qspTab && qspTab === TASK_TAB && !qspTaskId && (
-        <TaskItems ref={refetchRef} hideRelatedNode />
-      )}
-
-      {qspTab && qspTab === TASK_TAB && qspTaskId && (
-        <div>
-          <div className="flex bg-white text-sm">
-            <Link
-              to={constructPath(pathname, [
-                { name: QSP.TAB, value: TASK_TAB },
-                { name: QSP.TASK_ID, exclude: true },
-              ])}
-              className="flex items-center p-2 "
-            >
-              <Icon icon={"mdi:chevron-left"} />
-              All tasks
-            </Link>
-          </div>
-
-          <TaskItemDetails ref={refetchRef} />
-        </div>
+      {qspTab && (
+        <ObjectTabManager parentNodeSchema={schema} objectDetailsData={objectDetailsData} />
       )}
 
       <SlideOver
